Make `getdents` definition's type match the declaration.

+++ lib/libc/sys/getdents.c
@@ -1,5 +1,5 @@
 #include <dirent.h>
 
-ssize_t getdents(int fd, char *buf, size_t nbytes) {
+int getdents(int fd, char *buf, size_t nbytes) {
   return getdirentries(fd, buf, nbytes, NULL);
 }
